Issue 2336: Unit tests for equals() and hashCode() for concrete types
Implement the expectation that types are equal even when they are
separate/independent instances but have the same name.

Rationale is that name duplication isn't allowed in the same schema
anyway, so if the names are the same, then the two types coming from
the same schema must be equal.

Not implemented: checking that the types belong to the same schema.